requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
requests
pdfminer.six
pytesseract
Pillow
youtube-transcript-api
vosk
faiss-cpu
numpy
scikit-learn
nltk
sumy
pytest
flake8
radon
pydub
pdf2image
tqdm
sentence-transformers
